requirements.txt
Flask==2.0.1
mysql-connector-python==9.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3
